+++ lib/presentation/metrics_pill.dart
