+++ lib/presentation/notification_cards/src/constants.dart
