+++ lib/presentation/notification_cards/src/model/notification_card.dart
