+++ lib/presentation/notification_cards/src/stacked_cards/animated_offset_list.dart
