+++ lib/presentation/notification_cards/src/stacked_cards/expanded_list.dart
