+++ lib/presentation/transaction_stack.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter/src/widgets/container.dart';
 import 'package:flutter/src/widgets/framework.dart';
 // import 'package:stacked_notification_cards/stacked_notification_cards.dart';
-import 'package:flutter_annulus/presentation/home/notification_cards/stacked_notification_cards.dart';
+import 'package:flutter_annulus/presentation/notification_cards/stacked_notification_cards.dart';
 
 class TransactionStack extends StatefulWidget {
   final NotificationCard card;
